chore: move service functions from main to service.go
